run forge fmt

+++ contracts/src/Charitea.sol
@@ -4,49 +4,37 @@ pragma solidity ^0.8.27;
 contract Charitea {
     struct Fund {
         address owner;
-        uint value;
+        uint256 value;
         string data;
     }
 
-    uint fundCount;
+    uint256 fundCount;
     Fund[] public funds;
 
-    event FundCreated(
-        uint indexed fundIndex,
-        address indexed owner,
-        string data
-    );
-    event FundDonated(
-        uint indexed fundIndex,
-        address indexed donator,
-        uint amount
-    );
-    event FundWithdrawn(
-        uint indexed fundIndex,
-        address indexed withdrawer,
-        uint amount
-    );
+    event FundCreated(uint256 indexed fundIndex, address indexed owner, string data);
+    event FundDonated(uint256 indexed fundIndex, address indexed donator, uint256 amount);
+    event FundWithdrawn(uint256 indexed fundIndex, address indexed withdrawer, uint256 amount);
 
     error UnauthorizedWithdrawal();
 
-    function createFund(string calldata data) public returns (uint fundIndex) {
+    function createFund(string calldata data) public returns (uint256 fundIndex) {
         fundIndex = fundCount++;
         funds.push(Fund({owner: msg.sender, value: 0, data: data}));
         emit FundCreated(fundIndex, msg.sender, data);
     }
 
-    function donateFund(uint fundIndex) public payable {
+    function donateFund(uint256 fundIndex) public payable {
         funds[fundIndex].value += msg.value;
 
         emit FundDonated(fundIndex, msg.sender, msg.value);
     }
 
-    function withdrawFund(uint fundIndex) public {
+    function withdrawFund(uint256 fundIndex) public {
         Fund storage fund = funds[fundIndex];
 
         require(msg.sender == fund.owner, UnauthorizedWithdrawal());
 
-        uint toSend = fund.value;
+        uint256 toSend = fund.value;
         fund.value = 0;
 
         payable(fund.owner).transfer(toSend);
