+++ contracts/test/Charitea.t.sol
@@ -6,21 +6,9 @@ import {Vm} from "forge-std/Vm.sol";
 import {Charitea} from "../src/Charitea.sol";
 
 contract ChariteaTest is Test {
-    event FundCreated(
-        uint indexed fundIndex,
-        address indexed owner,
-        string data
-    );
-    event FundDonated(
-        uint indexed fundIndex,
-        address indexed donator,
-        uint amount
-    );
-    event FundWithdrawn(
-        uint indexed fundIndex,
-        address indexed withdrawer,
-        uint amount
-    );
+    event FundCreated(uint256 indexed fundIndex, address indexed owner, string data);
+    event FundDonated(uint256 indexed fundIndex, address indexed donator, uint256 amount);
+    event FundWithdrawn(uint256 indexed fundIndex, address indexed withdrawer, uint256 amount);
 
     Charitea charitea;
 
@@ -32,11 +20,9 @@ contract ChariteaTest is Test {
         vm.prank(msg.sender);
         vm.expectEmit(true, true, false, true);
         emit FundCreated(0, msg.sender, "test");
-        uint fundIndex = charitea.createFund("test");
+        uint256 fundIndex = charitea.createFund("test");
 
-        (address owner, uint value, string memory data) = charitea.funds(
-            fundIndex
-        );
+        (address owner, uint256 value, string memory data) = charitea.funds(fundIndex);
 
         assertEq(owner, msg.sender);
         assertEq(value, 0);
@@ -44,21 +30,21 @@ contract ChariteaTest is Test {
     }
 
     function test_DonateFund() public {
-        uint fundIndex = charitea.createFund("test");
+        uint256 fundIndex = charitea.createFund("test");
 
         vm.prank(msg.sender);
         vm.expectEmit(true, true, false, true);
         emit FundDonated(fundIndex, msg.sender, 10);
         charitea.donateFund{value: 10}(fundIndex);
 
-        (, uint value, ) = charitea.funds(fundIndex);
+        (, uint256 value,) = charitea.funds(fundIndex);
         assertEq(value, 10);
     }
 
     function test_WithdrawFund() public {
         Vm.Wallet memory wallet = vm.createWallet("fundOwner");
         vm.prank(wallet.addr);
-        uint fundIndex = charitea.createFund("test");
+        uint256 fundIndex = charitea.createFund("test");
         charitea.donateFund{value: 10}(fundIndex);
 
         vm.expectRevert(Charitea.UnauthorizedWithdrawal.selector);
@@ -68,7 +54,7 @@ contract ChariteaTest is Test {
         vm.expectEmit(true, true, false, true);
         emit FundWithdrawn(fundIndex, wallet.addr, 10);
         charitea.withdrawFund(fundIndex);
-        (, uint value, ) = charitea.funds(fundIndex);
+        (, uint256 value,) = charitea.funds(fundIndex);
         assertEq(value, 0);
         assertEq(wallet.addr.balance, 10);
     }
